requirements.txt
supabase
fastapi
python-dotenv
typing
uvicorn